+++ pyemma/coordinates/tests/test_tica.py
@@ -280,6 +280,8 @@ class TestTICAExtensive(unittest.TestCase):
     def test_describe(self):
         desc = self.tica_obj.describe()
         assert types.is_string(desc) or types.is_list_of_string(desc)
+        # describe on empty estimator
+        tica(lag=1).describe()
 
     def test_dimension(self):
         assert types.is_int(self.tica_obj.dimension())
@@ -292,6 +294,11 @@ class TestTICAExtensive(unittest.TestCase):
         assert tica.dimension() == 1
         with self.assertRaises(ValueError):  # trying to set both dim and subspace_variance is forbidden
             api.tica(data=self.X, lag=self.lag, dim=1, var_cutoff=0.9)
+
+        with self.assertRaises(ValueError):
+            api.tica(lag=self.lag, var_cutoff=0)
+        with self.assertRaises(ValueError):
+            api.tica(lag=self.lag, var_cutoff=1.1)
 
     def test_eigenvalues(self):
         eval = self.tica_obj.eigenvalues
@@ -365,7 +372,7 @@ class TestTICAExtensive(unittest.TestCase):
         self.tica_obj.number_of_trajectories() == 1
 
     def test_output_type(self):
-        assert self.tica_obj.output_type() == np.float32
+        assert self.tica_obj.output_type() == np.float32()
 
     def test_trajectory_length(self):
         assert self.tica_obj.trajectory_length(0) == self.T