+++ pyemma/datasets/api.py
@@ -120,26 +120,40 @@ def get_multi_temperature_data(kt0=1.0, kt1=5.0, length0=10000, length1=10000, n
     return mt_data
 
 
-def get_quadwell_data(nstep=10000, x0=0., nskip=1, dt=0.01, kT=10.0, mass=1.0, damping=1.0):
+def get_quadwell_data(ntraj=10, nstep=10000, x0=0., nskip=1, dt=0.001, kT=1.0, mass=1.0, damping=1.0):
     r""" Performs a Brownian dynamics simulation in the Prinz potential (quad well).
 
     Parameters
     ----------
+    ntraj: int, default=10
+        how many realizations will be computed
     nstep: int, default=10000
         number of time steps
     x0: float, default 0
         starting point for sampling
     nskip: int, default=1
         number of integrator steps
-    dt: float, default=0.01
+    dt: float, default=0.001
         time step size
-    kT: float, default=10.0
+    kT: float, default=1.0
         temperature factor
-    mass: float, default=1
+    mass: float, default=1.0
         mass
-    damping: float, default=1
+    damping: float, default=1.0
         damping factor of integrator
+
+    Returns
+    -------
+    trajectories : list of ndarray
+        realizations of the the brownian diffusion in the quadwell potential.
     """
     from .potentials import PrinzModel
     pw = PrinzModel(dt, kT, mass=mass, damping=damping)
-    return pw.sample(x0, nstep, nskip=nskip)
+    import warnings
+    import numpy as np
+    with warnings.catch_warnings(record=True) as w:
+        trajs = [pw.sample(x0, nstep, nskip=nskip) for _ in range(ntraj)]
+        if not np.all(tuple(np.isfinite(x) for x in trajs)):
+            raise RuntimeError('integrator detected invalid values in output. If you used a high temperature value (kT),'
+                               ' try decreasing the integration time step dt.')
+    return trajs