+++ pyemma/msm/estimators/implied_timescales.py
@@ -51,9 +51,8 @@ def _generate_lags(maxlag, multiplier):
 
     """
     # determine lag times
-    lags = []
+    lags = [1]
     # build default lag list
-    lags.append(1)
     lag = 1.0
     import decimal
     while lag <= maxlag:
@@ -64,6 +63,9 @@ def _generate_lags(maxlag, multiplier):
         if lag <= maxlag:
             ilag = int(lag)
             lags.append(ilag)
+    # always include the maximal requested lag time.
+    if maxlag not in lags:
+        lags.append(maxlag)
     return np.array(lags)
 
 
@@ -121,6 +123,8 @@ class ImpliedTimescales(Estimator, NJobsMixIn, SerializableMixIn):
         self._its = None
         # sampled its's. 3D-array with indexing: lagtime, its, sample
         self._its_samples = None
+        # fingerprint of the last estimate input.
+        self._last_dtrajs_input_hash = None
 
     def estimate(self, X, **params):
         """
@@ -156,9 +160,11 @@ class ImpliedTimescales(Estimator, NJobsMixIn, SerializableMixIn):
         if self._estimated:
             # if dtrajs has now changed, unset the _estimated flag to re-set every derived quantity.
             assert hasattr(self, '_last_dtrajs_input_hash')
-            if self._last_dtrajs_input_hash != _hash_dtrajs(dtrajs):
+            current_hash = _hash_dtrajs(dtrajs)
+            if self._last_dtrajs_input_hash != current_hash:
                 self.logger.warning("estimating from new data, discard all previously computed models.")
                 self._estimated = False
+                self._last_dtrajs_input_hash = current_hash
         else:
             self._last_dtrajs_input_hash = _hash_dtrajs(dtrajs)